src/chatbot_backend/requirements.txt
cachetools==6.1.0
fastapi==0.115.14
ipython==8.18.1
langchain-core==0.3.68
langchain-openai==0.3.27
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-checkpoint-sqlite==2.0.10
numpy==2.0.2
pandas==2.3.0
pandasql==0.7.3
sqlparse==0.5.3
uvicorn==0.35.0